+++ HISTORY.md
@@ -77,6 +77,7 @@
 * gui: removed redundant "duplicate costume" user event
 * gui: added costume # to "remove costume" user event
 * gui: added "rename sound" user event
+* gui: added "delete sound" user event
 
 ### 2022-12-01
 * blocks: don't show "result pic" option in context menu with "click-to-run" disabled